RRSP Solver Documentation

The purpose of this document is to illustrate the user API of Robust Recoverable Shortest Path Solver. Technical documentation of the solver is left in the source code in the form of docstrings and inline comments.

The documentation adheres to the 0.1.0 version of Rrsp package.

Contents

RRSP Solver Documentation

Contents

How to install

API Data Types

Setting the Optimizer

Parsing Input

Solving Shortest Path Problems

Working Example

Index

How to install

Use Pkg package manager to install this package for a session.

Clone the repository.

Add the package to local registry by

julia

import Pkg; Pkg.develop(path="./Rrsp")

Build the package by importing it for the first time:

import Rrsp

Dependencies

The only package dependency of Rrsp is JuMP. Users are responsible for installing MIP solvers they want to use with Rrsp.

API Data Types

For simplicity, the API uses only one object type for input and output, respectively.

Rrsp.RrspInstance — Type

struct RrspInstance
    graph::Graph
    s_idx::Integer
    t_idx::Integer
    neighbourhood::NeighbourhoodType
    k::Integer
    gamma::Float64
end

The structure containing serialized RRSP problem instance. It can be used as an opaque type, since the solving functions only input parameter is the instance itself.

Rrsp.RrspSolution — Type

struct RrspSolution
    first_stage_path::RrspPath
    second_stage_path::RrspPath
    value::Float64
end

The structure containing optimal pair of paths and the optimal value of the solution to the RRSP problems.

Every solver API function returns RrspSolution structure object. When solution is infeasible, the RrspSolution structure is returned empty – the paths structures are initialized to $\mathbf{0}$ vectors and value is Inf.

For convenience, the same structure is used when returning the solution for RRSP subproblems, which concern only one optimal path. In this case, as to which path of the set is given by function docstring.

The RrspPath structure, the building block of the output object type, is left transparent:

Rrsp.RrspPath — Type

struct RrspPath
    arcs::Vector{Bool}
end

The structure for encapsulating a path of the problem instance graph. Paths are represented by characteristic vectors of the arc set of the instance graph as this representation allows for easier constraints formulation in MIP modeling – the principal aspect of the solver.

To fetch the arcs of which the solution path consist you can use the RrspPath.arcs vector.

Setting the Optimizer

Before using the solver you need to supply the optimizer of your choice. This is done by simple assignment to the global variable Rrsp.optimizer. For example, if we want to use Cbc optimizer to solve MIP models we must have assigned

Rrsp.optimizer = Cbc.Optimizer

somewhere before using Rrsp API. The optimizer can be set dynamically between calls to the API, for example to test solving times using different optimizers.

Parsing Input

The instance of RRSP is assumed to be given in a text file. Parsing the instance is separate from solving the problem. Hence, you can keep multiple RrspInstance objects at runtime at once.

Rrsp.parseInstanceFromFile — Function

parseInstanceFromFile(file_name::String)::RrspInstance

Return serialized RRSP instance given in file_name. The input file should conform to the following format:

s t N k G
<arc>
...
<arc>

where

s – path start node id,

t – path end node id,

N – neighbourhood type; one of INC, EXC, SYM_DIFF,

k – neighbourhood size parameter,

G – budget,

and <arc> is a line serializing one arc:

n1 n2 c1 c2 d

where

n1 – arc start node id,

n2 – arc end node id,

c1 – first stage cost,

c2 – second stage cost lower bound,

d – second stage cost maximum deviation.

Example:

1 3 INC 1 100.0
1 2 0.0 50.0 10.0
1 2 0.0 0.0 100.0
2 3 0.0 20.0 30.0
2 3 0.0 40.0 0.0

Solving Shortest Path Problems

Once the instance is deserialized you can choose which problem should be solved with it. Solving the instance doesn't change its data so one instance object can be resused to obtain solutions to different variants of the problem.

Functions solving RRSP and the subproblems are grouped here according to the problem type:

Deterministic SP

Rrsp.solveDeterministicShortestPath — Function

solveDeterministicShortestPath(instance::RrspInstance)::RrspSolution

Returns a shortest $s-t$ path in the graph instance.graph with respect to Cost.first costs, where $s$ and $t$ are nodes given by instance.s_idx, instance.t_idx indices in instance.graph.arcs array.

The path is stored in RrspSolution.first_stage_path of the returned structure and its cost is assigned to RrspSolution.value.

The path is computed using LP model.

Incremental SP

Rrsp.solveIncrementalShortestPath — Function

solveIncrementalShortestPath(instance::RrspInstance, x::Path)::RrspSolution

Returns a shortest $s-t$ path in the graph instance.graph with respect to Cost.second costs, where $s$ and $t$ are nodes given by instance.s_idx, instance.t_idx indices in instance.graph.arcs array.

If the costs for incremental problem are different than second stage costs lower bounds, they should be set before calling the function.

The path is in the neighbourhood of the path x; the neighbourhood type is selected with instance.neighbourhood and the neighbourhood size with instance.k values.

The path is stored in RrspSolution.second_stage_path of the returned structure and its cost is assigned to RrspSolution.value. For integrality of the solution, the input parameter path x is stored in RrspSolution.first_stage_path. However, the first stage cost doesn't affect the value of the solution.

The path is computed using compact MIP model.

Recoverable SP

Rrsp.solveRecoverableShortestPath — Function

solveRecoverableShortestPath(instance::RrspInstance)::RrspSolution

Returns optimal first and second stage shortest $s-t$ path in the graph instance.graph for Recoverable Shortest Path, to which the RRSP problem with interval uncertainty reduces to. The nodes $s$ and $t$ are given by instance.s_idx, instance.t_idx indices in instance.graph.arcs array.

The paths are computed using compact MIP model.

Rrsp.solveRecoverableShortestPathInAsp — Function

solveRecoverableShortestPathInAsp(instance::RrspInstance)::RrspSolution

Returns optimal first and second stage shortest $s-t$ path in the graph instance.graph for Recoverable Shortest Path in Arc Series-Parallel (ASP) graph.

The instance.graph is assumed to be ASP. Hence, it's not validated further. If this is not the case, then the behaviour and the return value of this function are undefined.

The paths are computed using combinatorial algoritm.

NOTE: The only supported neighbourhood is inclusion for now!

Recoverable Robust SP

Rrsp.solveRrspContBudget — Function

solveRrspContBudget(instance::RrspInstance)::RrspSolution

Returns optimal first and second stage shortest $s-t$ path in the graph instance.graph for Recoverable Robust Shortest Path with continuous budgeted uncertainty. The nodes $s$ and $t$ are given by instance.s_idx, instance.t_idx indices in instance.graph.arcs array. The maximum budget value is equal to instance.gamma.

If the graph is a DAG, then use solveRrspContBudgetDag function.

The paths are computed using compact MIP model.

Rrsp.solveRrspContBudgetDag — Function

solveRrspContBudgetDag(instance::RrspInstance)::RrspSolution

Returns optimal first and second stage shortest $s-t$ path in the graph instance.graph for Recoverable Robust Shortest Path with continuous budgeted uncertainty when instance.graph is a directed acyclic graph. The nodes $s$ and $t$ are given by instance.s_idx, instance.t_idx indices in instance.graph.arcs array. The maximum budget value is equal to instance.gamma.

The paths are computed by solving a family of $m + 1$ compact MIP models, where $m$ is the number of arcs in the DAG. Since the graph is acyclic, the model does not contain anti-cyclic constraints, unlike the model for general digraphs.

Working Example

To illustrate the use of API we present an example of solving a small problem instance. In this example we use Cbc for solving MIP models.

Assume that there exists a file under data/instance.rrsp path containing a problem instance written down as shown in Parsing Input section.
Make sure Rrsp definitions are visible. Use import to avoid name collisions in your project. If not an issue, you can use using.

julia> import Rrsp

Import Cbc optimizer package and use it to set Rrsp optimizer (you can choose any JuMP conforming optimizer able to solve MIP models):

julia> import Cbc

julia> Rrsp.optimizer = Cbc.Optimizer

Parse the file to create an instance object:

julia> instance::Rrsp.RrspInstance = Rrsp.parseInstanceFromFile("data/instance.rrsp")

Use one of the solver functions to solve specific variant of the RRSP problem. Here, we use function for solving RRSP with continuous budget in general digraphs:

solution::Rrsp.RrspSolution = Rrsp.solveRrspContBudget(instance)

You can check if solving was succesful with solution.value != Inf. To access the optimal paths use solution.first_stage_path and solution.second_stage_path objects.

Index

Rrsp.RrspInstance

Rrsp.RrspPath

Rrsp.RrspSolution

Rrsp.parseInstanceFromFile

Rrsp.solveDeterministicShortestPath

Rrsp.solveIncrementalShortestPath

Rrsp.solveRecoverableShortestPath

Rrsp.solveRecoverableShortestPathInAsp

Rrsp.solveRrspContBudget

Rrsp.solveRrspContBudgetDag